refactor: Workbench.tsx logic split into hooks to fix ESLint max-lines rules (#1207)
